requirements.txt
discord.py==1.2.2
discord==1.0.1
python-dotenv==0.10.3
